Implement CR fixes
